requirements.txt
langchain>=0.1.0
tavily-python>=0.1.0
alpha_vantage>=2.3.1
yfinance>=0.2.28
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.0.0
